% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_toml.r
\name{encode_toml}
\alias{encode_toml}
\alias{encode_toml.list}
\alias{encode_toml.default}
\alias{encode_toml.character}
\alias{encode_toml.integer}
\alias{encode_toml.logical}
\alias{encode_toml.numeric}
\alias{encode_toml.Date}
\alias{encode_toml.POSIXt}
\alias{encode_toml.datetimeoffset}
\title{Encode R object to TOML}
\usage{
encode_toml(x, ...)

\method{encode_toml}{list}(x, ..., prefix = NULL, inline = FALSE)

\method{encode_toml}{default}(x, ...)

\method{encode_toml}{character}(x, ...)

\method{encode_toml}{integer}(x, ..., big.mark = "_", big.interval = 3L)

\method{encode_toml}{logical}(x, ...)

\method{encode_toml}{numeric}(
  x,
  ...,
  big.mark = "_",
  big.interval = 3L,
  small.mark = big.mark,
  small.interval = big.interval
)

\method{encode_toml}{Date}(x, ...)

\method{encode_toml}{POSIXt}(x, ...)

\method{encode_toml}{datetimeoffset}(x, ...)
}
\arguments{
\item{x}{R object to encode}

\item{...}{Currently ignored}

\item{prefix}{Prefix for keys.
If \code{NULL} and \code{isFALSE(inline)} assume this is the top-level aka root table.}

\item{inline}{Encode as a TOML "inline" table.}

\item{big.mark, big.interval}{\code{big.mark} is used as a mark between every \code{big.interval} decimals
(before the decimal point).
TOML allows this to be either \code{"_"} or \code{""}.}

\item{small.mark, small.interval}{\code{small.mark} is used as a mark between every \code{small.interval} decimals
after the decimal point.
TOML allows this to be either \code{"_"} or \code{""}.}
}
\value{
A character vector
}
\description{
\code{encode_toml()} is an S3 method that encodes R object to TOML
}
