% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_toml.r
\name{write_toml}
\alias{write_toml}
\title{Write TOML config files}
\usage{
write_toml(x, con = stdout(), ...)
}
\arguments{
\item{x}{An R object with an \code{\link[=encode_toml]{encode_toml()}} method.}

\item{con}{A connection object or a character string.
Passed to \code{\link[base:writeLines]{base::writeLines()}}.}

\item{...}{Passed to \code{\link[=encode_toml]{encode_toml()}}.}
}
\description{
\code{write_toml()} encodes R objects into TOML
and then writes them to a connection.
}
