% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_toml.r
\name{encode_toml}
\alias{encode_toml}
\alias{encode_toml.list}
\alias{encode_toml.default}
\alias{encode_toml.character}
\alias{encode_toml.integer}
\alias{encode_toml.logical}
\alias{encode_toml.numeric}
\alias{encode_toml.Date}
\alias{encode_toml.POSIXt}
\alias{encode_toml.datetimeoffset}
\title{Encode R object to TOML}
\usage{
encode_toml(x, ...)

\method{encode_toml}{list}(x, ..., prefix = NULL, inline = FALSE)

\method{encode_toml}{default}(x, ...)

\method{encode_toml}{character}(x, ...)

\method{encode_toml}{integer}(x, ...)

\method{encode_toml}{logical}(x, ...)

\method{encode_toml}{numeric}(x, ...)

\method{encode_toml}{Date}(x, ...)

\method{encode_toml}{POSIXt}(x, ...)

\method{encode_toml}{datetimeoffset}(x, ...)
}
\arguments{
\item{x}{R object to encode}

\item{...}{Currently ignored}

\item{prefix}{Prefix for keys.  If \code{NULL} assume this is the top-level aka root table.}

\item{inline}{Encode as a TOML "inline" table.}
}
\value{
A character vector
}
\description{
\code{encode_toml()} is an S3 method that encodes R object to TOML
}
